refactor select pokemon type
