+++ src/features/select-pokemon-type/hooks/index.ts
@@ -1,0 +1,2 @@
+export { default as useSelectPokemonType } from "./useSelectPokemonType";
+export { default as usePokemonTypes } from "./usePokemonTypes";
